ref(symcache): Make symbolic-symcache dependent on symbolic-il2cpp instead of the other way around
