+++ examples/symcache_debug/src/main.rs
@@ -10,6 +10,7 @@ use symbolic::common::{Arch, ByteView, DSymPathExt, Language, SelfCell};
 use symbolic::debuginfo::macho::BcSymbolMap;
 use symbolic::debuginfo::Archive;
 use symbolic::demangle::{Demangle, DemangleOptions};
+#[cfg(feature = "il2cpp")]
 use symbolic::il2cpp::LineMapping;
 use symbolic::symcache::transform::{self, Transformer};
 use symbolic::symcache::{SymCache, SymCacheWriter};
@@ -84,11 +85,14 @@ fn execute(matches: &ArgMatches) -> Result<()> {
             writer.add_transformer(bcsymbolmap);
         }
 
-        if let Some(linemapping_file) = matches.value_of("linemapping_file") {
-            let linemapping_path = Path::new(linemapping_file);
-            let linemapping_buffer = ByteView::open(linemapping_path)?;
-            if let Some(linemapping) = LineMapping::parse(&linemapping_buffer) {
-                writer.add_transformer(linemapping);
+        #[cfg(feature = "il2cpp")]
+        {
+            if let Some(linemapping_file) = matches.value_of("linemapping_file") {
+                let linemapping_path = Path::new(linemapping_file);
+                let linemapping_buffer = ByteView::open(linemapping_path)?;
+                if let Some(linemapping) = LineMapping::parse(&linemapping_buffer) {
+                    writer.add_transformer(linemapping);
+                }
             }
         }
