+++ symbolic-il2cpp/Cargo.toml
@@ -20,7 +20,6 @@ scroll = "0.11"
 serde_json = "1.0.79"
 symbolic-common = { version = "8.6.1", path = "../symbolic-common" }
 symbolic-debuginfo = { version = "8.6.1", path = "../symbolic-debuginfo" }
-symbolic-symcache = { version = "8.6.1", path = "../symbolic-symcache" }
 thiserror = "1.0.20"
 
 [dev-dependencies]
